Update to 1.21.3 and other updates

+++ mednafen/gba/thumb.cpp
@@ -1852,7 +1852,7 @@ switch(opcode >> 8) {
  case 0xd0:
    // BEQ offset
    if(Z_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1863,7 +1863,7 @@ switch(opcode >> 8) {
  case 0xd1:
    // BNE offset
    if(!Z_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1874,7 +1874,7 @@ switch(opcode >> 8) {
  case 0xd2:
    // BCS offset
    if(C_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1885,7 +1885,7 @@ switch(opcode >> 8) {
  case 0xd3:
    // BCC offset
    if(!C_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1896,7 +1896,7 @@ switch(opcode >> 8) {
  case 0xd4:
    // BMI offset
    if(N_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1907,7 +1907,7 @@ switch(opcode >> 8) {
  case 0xd5:
    // BPL offset
    if(!N_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1918,7 +1918,7 @@ switch(opcode >> 8) {
  case 0xd6:
    // BVS offset
    if(V_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1929,7 +1929,7 @@ switch(opcode >> 8) {
  case 0xd7:
    // BVC offset
    if(!V_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1940,7 +1940,7 @@ switch(opcode >> 8) {
  case 0xd8:
    // BHI offset
    if(C_FLAG && !Z_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1951,7 +1951,7 @@ switch(opcode >> 8) {
  case 0xd9:
    // BLS offset
    if(!C_FLAG || Z_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1962,7 +1962,7 @@ switch(opcode >> 8) {
  case 0xda:
    // BGE offset
    if(N_FLAG == V_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1973,7 +1973,7 @@ switch(opcode >> 8) {
  case 0xdb:
    // BLT offset
    if(N_FLAG != V_FLAG) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1984,7 +1984,7 @@ switch(opcode >> 8) {
  case 0xdc:
    // BGT offset
    if(!Z_FLAG && (N_FLAG == V_FLAG)) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
@@ -1995,7 +1995,7 @@ switch(opcode >> 8) {
  case 0xdd:
    // BLE offset
    if(Z_FLAG || (N_FLAG != V_FLAG)) {
-     reg[15].I += ((int8)(opcode & 0xFF)) << 1;       
+     reg[15].I += (uint32)((int8)(opcode & 0xFF)) << 1;
      armNextPC = reg[15].I;
      reg[15].I += 2;
      THUMB_PREFETCH;
